+++ src/test/java/com/caiata/web/TestIlMeteo.java
@@ -1,17 +1,20 @@
 package com.caiata.web;
 
-import com.caiata.IlMeteoSteps;
+import com.caiata.steps.IlMeteoSteps;
 import com.caiata.ManagementDriver;
 import com.caiata.Utility;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
+import org.openqa.selenium.*;
+import org.openqa.selenium.support.ui.FluentWait;
+import org.openqa.selenium.support.ui.Wait;
+import org.openqa.selenium.support.ui.WebDriverWait;
 
+import java.time.Duration;
 import java.util.Properties;
+import java.util.function.Function;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
@@ -30,6 +33,8 @@ public class TestIlMeteo{
             ManagementDriver.startDriver();
             driver = ManagementDriver.getDriver();
             steps = new IlMeteoSteps();
+           // webElement = new WebDriverWait(driver, Duration.ofSeconds(3)).until(driver -> driver.findElement(By.id(("tab" + 1))));
+
         }
 
         @BeforeEach
@@ -39,11 +44,11 @@ public class TestIlMeteo{
 
         @Order(1)
         @ParameterizedTest(name = "q = {0}")
-        @ValueSource(strings = {"Picerno"})
+        @ValueSource(strings = {"Picerno", "ciao"})
         @DisplayName("controllare esistenza prodotto ricercato")
         void test_001(String q) {
             driver.get(webProp.getProperty("ilmeteo.url"));
-            steps.closeBanner(webProp);
+            steps.closeBannerFW(webProp);
             steps.search(webProp, q);
             String result = driver.findElement(By.id(webProp.getProperty("id.page.title"))).getText();
             System.out.println(result);
@@ -61,12 +66,12 @@ public class TestIlMeteo{
             driver.get(webProp.getProperty("ilmeteo.url"));
             try{
                 for(int i = 1; i <= 14; i++){
-                    webElement = driver.findElement(By.id(("tab" + 1)));
+                    webElement = new WebDriverWait(driver, Duration.ofSeconds(3)).until(driver -> driver.findElement(By.id(("tab" + 1))));
+                    //webElement = driver.findElement(By.id(("tab" + 1)));
                     nome = webElement.getText();
                     webElement.click();
-                    Thread.sleep(500);
-                    if(nome.equals("Home")){
-                        continue;
+                    //Thread.sleep(500);
+                    if(!nome.equals("Home")){
                     }else{
                         if(!driver.findElement(By.id(webProp.getProperty("id.page.title"))).getText().toLowerCase().contains(nome.toLowerCase())){
                             System.out.println(nome + " Non trovato!");
@@ -76,7 +81,7 @@ public class TestIlMeteo{
                         assertTrue(driver.findElement(By.id(webProp.getProperty("id.page.title"))).getText().toLowerCase().contains(nome.toLowerCase()));
                     }
                 }
-            }catch(InterruptedException e){
+            }catch(NoSuchElementException | TimeoutException e){
                 System.out.println("Banner non trovato !");
             }
         }
@@ -89,15 +94,39 @@ public class TestIlMeteo{
         String nome;
         driver.get(webProp.getProperty("ilmeteo.url"));
         try {
-            webElement = driver.findElement(By.id((tab)));
+            webElement = new WebDriverWait(driver, Duration.ofSeconds(3)).until(driver -> driver.findElement(By.id((tab))));
             nome = webElement.getText();
             webElement.click();
-            Thread.sleep(500);
             assertTrue(driver.findElement(By.id(webProp.getProperty("id.page.title"))).isEnabled());
             if (!nome.equals("Home")) {
                 assertTrue(driver.findElement(By.id(webProp.getProperty("id.page.title"))).getText().toLowerCase().contains(nome.toLowerCase()));
             }
-            }catch (InterruptedException e) {
+            }catch (NoSuchElementException | TimeoutException e) {
+            System.out.println("Banner non Trovato");
+        }
+    }
+
+    @ParameterizedTest(name = "{0}")
+    @DisplayName("controllo titoli con csv")
+    @CsvSource({"tab1","tab2","tab3","tab4","tab5","tab6","tab7","tab8","tab9","tab10","tab11","tab12","tab13","tab14"})
+    @Order(3)
+    void test_003_ControlloFW(String tab){
+        String nome;
+        driver.get(webProp.getProperty("ilmeteo.url"));
+        try {
+            Wait<WebDriver> wait = new FluentWait<>(driver)
+                    .withTimeout(Duration.ofSeconds(3))
+                    .pollingEvery(Duration.ofSeconds(3))
+                    .ignoring(NoSuchElementException.class);
+            webElement = wait.until(driver -> driver.findElement(By.id("id.page.title")));
+
+            nome = webElement.getText();
+            webElement.click();
+            assertTrue(driver.findElement(By.id(webProp.getProperty("id.page.title"))).isEnabled());
+            if (!nome.equals("Home")) {
+                assertTrue(driver.findElement(By.id(webProp.getProperty("id.page.title"))).getText().toLowerCase().contains(nome.toLowerCase()));
+            }
+        }catch (NoSuchElementException | TimeoutException e) {
             System.out.println("Banner non Trovato");
         }
     }
@@ -107,6 +136,21 @@ public class TestIlMeteo{
     @DisplayName("controllo titoli con lista")
     void test_004() throws InterruptedException {
         driver.get(webProp.getProperty("ilmeteo.url"));
+        String handle = driver.getWindowHandle();
+        for(WebElement element : steps.getMenuTabs(webProp)){
+            String elementText = element.getText();
+            if(!elementText.equals("Home")){
+                String a = element.getAttribute("href");
+                driver.switchTo().newWindow(WindowType.TAB);
+                driver.get(a);
+                //webElement = new WebDriverWait(driver, Duration.ofSeconds(3)).until(driver -> driver.findElement(By.id(webProp.getProperty("id.page.title"))));
+                Thread.sleep(500);
+                assertTrue(driver.findElement(By.id(webProp.getProperty("id.page.title"))).getText().toLowerCase().contains(elementText.toLowerCase()));
+                driver.close();
+                driver.switchTo().window(handle);
+            }
+        }
+        /*
         for(int i = 2; i < steps.getMenuTabs(webProp).size(); i++){
             webElement = steps.getMenuTabs(webProp).get(i);
             webElement = driver.findElement(By.id("tab"+i));
@@ -119,7 +163,7 @@ public class TestIlMeteo{
 
             System.out.println("Titolo superiore : " + webElement.getText());
             System.out.println("Titolo stampato sotto : " + tmp.getText());
-        }
+        }*/
     }
 
     @AfterEach