+++ CHANGELOG.md
@@ -10,3 +10,6 @@ _`**VERSIONE 1.0**`_
 - Aggiunto `meteoSteps e TestMeteo`
 - Aggiunto `Test per i titoli de ILMETEO`
 - Aggiunto `Test per i titoli de ILMETEO con liste`
+- Aggiunto `Test per le categorie Ebay`
+- Aggiunto `Test per la selezione delle categorie Ebay`
+- Aggiunti `fluent wait e webDriverWait`